@@ -1,13 +1,16 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
-import AuthContext from './Context/auth-context';
+import React from "react";
+import { createRoot } from "react-dom/client";
+import "./index.css";
+import App from "./App";
+import * as serviceWorker from "./serviceWorker";
+import AuthContext from "./Context/auth-context";
 
-ReactDOM.render(
-    <AuthContext>
-        <App />
-    </AuthContext>, document.getElementById('root'));
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
+  <AuthContext>
+    <App />
+  </AuthContext>
+);
 
 serviceWorker.unregister();
